update `evaluate` default shortcuts
- update `evaluate` default shortcut to `Alt + D` (#590)
- update `evaluate-translate` default shortcut to `Alt + T` (#590)
