+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.19
+
+- update `evaluate` default shortcut to `Alt + D` (#590)
+- update `evaluate-translate` default shortcut to `Alt + T` (#590)
+- fix poe overlay not visible after focus change (#587)
+
 ## 0.6.18 (2020-04-08)
 
 - add file cache for all requests to further reduce total request count
