+++ src/app/layout/component/index.ts
@@ -1,3 +1,4 @@
-export * from './user-settings-dialog/user-settings-dialog.component';
 export * from './user-settings-feature-container/user-settings-feature-container.component';
 export * from './user-settings-form/user-settings-form.component';
+export * from './user-settings-help/user-settings-help.component';
+
