--- src/app/layout/component/user-settings-dialog/user-settings-dialog.component.html
@@ -1,20 +1,0 @@
-<ng-container *ngIf="settings && features">
-    <mat-tab-group mat-stretch-tabs animationDuration="0ms" (selectedIndexChange)="onSelectedIndexChange($event)">        
-        <mat-tab [label]="'settings.general' | translate">
-            <div class="tab-content">
-                <app-user-settings-form [settings]="settings"></app-user-settings-form>
-            </div>
-        </mat-tab>
-        <mat-tab [label]="feature.name | translate" *ngFor="let feature of features">
-            <div class="tab-content">
-                <app-user-settings-feature-container [component]="feature.component" [settings]="settings">
-                </app-user-settings-feature-container>
-            </div>
-        </mat-tab>
-        <mat-tab [label]="'help.name' | translate">
-            <div class="tab-content">
-                <app-user-settings-help></app-user-settings-help>
-            </div>
-        </mat-tab>
-    </mat-tab-group>
-</ng-container>
