--- src/app/layout/component/user-settings-dialog/user-settings-dialog.component.scss
@@ -1,3 +1,0 @@
-.tab-content {
-  padding: 12px 12px 0px 12px;
-}
