+++ src/app/layout/page/index.ts
@@ -1,0 +1,2 @@
+export * from './overlay/overlay.component'
+export * from './user-settings/user-settings.component'
