+++ src/app/layout/page/user-settings/user-settings.component.scss
@@ -3,3 +3,7 @@
         background: #1c1c1c;
     }
 }
+
+.tab-content {
+    padding: 12px 12px 0px 12px;
+}
